requirements.txt
albumentations==1.2.0
anykeystore==0.2
apex==0.9.10.dev0
certifi==2019.11.28
chardet==3.0.4
cryptacular==1.6.2
cycler==0.11.0
dbus-python==1.2.16
defusedxml==0.7.1
fonttools==4.33.3
greenlet==1.1.2
hupper==1.10.3
idna==2.8
imageio==2.19.3
joblib==1.1.0
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
networkx==2.8.4
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.5.0
PasteDeploy==2.1.1
pbkdf2==1.3
Pillow==9.1.1
plaster==1.0
plaster-pastedeploy==0.7
PyGObject==3.36.0
pyparsing==3.0.9
pyramid==2.0
pyramid-mailer==0.15.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
repoze.sendmail==4.4.1
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
scikit-image==0.18.3
scikit-learn==1.1.1
scipy==1.8.1
six==1.14.0
SQLAlchemy==1.4.39
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.5.4
torch==1.11.0+cu113
torchvision==0.12.0+cu113
transaction==3.0.1
translationstring==1.4
typing-extensions==4.2.0
urllib3==1.25.8
velruse==1.1.1
venusian==3.0.0
WebOb==1.8.7
WTForms==3.0.1
wtforms-recaptcha==0.3.2
yacs==0.1.8
zope.deprecation==4.4.0
zope.interface==5.4.0
zope.sqlalchemy==1.6
